wow helpful comments
wow so documentation
very clear
much helpful
also some todo

</shibe>

To be honest, the doge code isn't all that clear and concise. It was written
quickly for teh lelz and as a result it isn't of stellar quality. It does what
it should, but it is slightly hairy. I've gotten confused by it so much that
I thought I might add some actual comments.

This is as much for me as it is for anyone else.