wow show all process
wow list such clumsy
many miss
much process

</doge>

The known processes list will never really be helpful since it's kind of
opt-in. And, quite frankly, the system processes are also a part of the shibe
of ones computer. So, from now on all of the processes will be shown.

Also increase number of processes to 2.

Closes #17 since there is no need to curate a list anymore.